refactor: Update Constants and InputDemo.jsx
